refactor: 승패 판단 로직 이동
