+++ docs/README.md
@@ -12,7 +12,6 @@
 - 플레이어 게임 결과(PlayerGameResult)
 - 돈(Money)
 - 배팅정보(Bettings)
-- 승패 판단(Judge)
 
 ## 블랙잭 승패 규칙
 
@@ -88,6 +87,8 @@
     - [x] 승은 1배의 배팅 수익률을 가진다.
     - [x] 블랙잭 승은 1.5배의 배팅 수익률을 가진다.
     - [x] 무는 0.0배의 배팅 수익률을 가진다.
+- [x] 플레이어와 딜러를 통해 플레이어의 승패를 판단할 수 있다.
+
 
 ### 돈
 
@@ -101,6 +102,3 @@
 - [x] 딜러 수익을 알 수 있다.
 - [x] 플레이어들의 수익을 알 수 있다.
 
-### 승패 판단
-
-- [x] 플레이어와 딜러를 통해 플레이어의 승패를 판단할 수 있다.
